convert material
