+++ Engine/include/sge.h
@@ -61,4 +61,5 @@
 #include "component/PhysicsComponent.h"
 #include "camera/CameraControllerFreeLook.h"
 #include "component/MeshComponent.h"
+#include "component/MaterialComponent.h"
 #include "runtime/Prefab.h"
