+++ Engine/src/component/Component.h
@@ -157,89 +157,6 @@ struct EngineAPI NativeScriptComponent : public Component
         static void attachToEntity(std::shared_ptr<Component>, Entity, Scene&);
 };
 
-//struct EngineAPI MeshComponent : public Component
-//{
-//	MeshComponent(const Resource<MeshCollection>& mesh) : mesh(mesh) {};
-//	MeshComponent() = default;
-//
-//	template <class Archive>
-//	void serialize(Archive& archive) {
-//		SERIALIZED_MEMBER(mesh);
-//		SERIALIZED_MEMBER(materialSlot);
-//
-//	}
-//
-//	float materialSlot = 0; // todo this will be used (probably as a list) to support multi material models
-//        Resource<MeshCollection> mesh = Resource<MeshCollection>::empty;
-//        static void attachToEntity(std::shared_ptr<Component>, Entity, Scene&);
-//};
-
-
-//struct EngineAPI MeshArrayRendererComponent : public Component
-//{
-//	MeshArrayRendererComponent() = default;
-//	
-//	void addMesh(const MeshRenderer& mesh)
-//	{
-//		meshes.push_back(mesh);
-//	}
-//
-//	auto begin() { return meshes.begin(); }
-//	auto end() { return meshes.end(); }
-//	auto begin() const { return meshes.begin(); }
-//	auto end() const { return meshes.end(); }
-//
-//	template <class Archive>
-//	void serialize(Archive& archive) {
-//		archive(meshes);
-//	}
-//
-//	std::vector<MeshRenderer> meshes;
-//};
-
-struct EngineAPI MaterialComponent : public Component
-{
-	MaterialComponent();
-	
-
-	auto begin() { return materials.begin(); }
-	auto end() { return materials.end(); }
-	auto begin() const { return materials.begin(); }
-	auto end() const { return materials.end(); }
-
-	void addMaterial(const std::shared_ptr<Material>& mat)
-	{
-		materials[count++] = mat;
-	}
-
-	void setMaterial(int index, const std::shared_ptr<Material>& mat)
-	{
-		materials[index] = mat;
-	}
-
-	std::shared_ptr<Material> at(int index)
-	{
-		auto iter = materials.find(index);
-		if (iter == materials.end())
-		{
-			return Engine::get()->getDefaultMaterial();
-		}
-		return iter->second;
-	}
-
-	static void attachToEntity(std::shared_ptr<Component>, Entity, Scene&);
-
-	template <class Archive>
-	void serialize(Archive& archive) {
-		SERIALIZED_MEMBER(materials);
-		SERIALIZED_MEMBER(count);
-	}
-
-    std::map<int, std::shared_ptr<Material>> materials;
-    int count = 0;
-    
-};
-
 
 struct EngineAPI ObjectComponent : public Component
 {
