+++ Engine/src/geometry/ShapeFactory.cpp
@@ -17,6 +17,7 @@
 #include "core/CacheSystem.h"
 #include "systems/BuiltInMeshes.h"
 #include "component/MeshComponent.h"
+#include "component/MaterialComponent.h"
 
 Entity ShapeFactory::createEntity(SGE_Regsitry* registry, const std::string& name)
 {
